Backend/OnlineVoting/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.12
amqp==5.2.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.10.3
beautifulsoup4==4.11.2
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2022.9.24
cffi==1.16.0
channels==4.0.0
chardet==5.1.0
charset-normalizer==2.1.1
clib==0.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.25.0
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cssselect2==0.7.0
cvzone==1.5.6
cycler==0.11.0
distlib==0.3.7
Django==4.2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-renderpdf==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dlib==19.24.0
dnspython==2.4.2
docutils==0.19
emoji==2.2.0
eventlet==0.33.3
face-recognition==1.3.0
face-recognition-models==0.3.0
ffpyplayer==4.3.5
filelock==3.13.1
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
Foundations==2.0.7
frozenlist==1.3.3
gpio==1.0.0
greenlet==3.0.1
html5lib==1.1
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
kombu==5.3.4
linecache2==1.0.0
Manager==2.0.3
MarkupSafe==2.1.1
matplotlib==3.6.3
mediapipe==0.9.0.1
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.23.4
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
ordereddict==1.1
packaging==21.3
pandas==1.5.1
Pillow==9.4.0
playsound==1.3.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psycopg==3.1.12
psycopg2==2.9.9
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.8.0
PyGetWindow==0.0.9
pyglet==2.0.3
Pygments==2.13.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pyphen==0.14.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.5
pywin32==305
redis==4.6.0
requests==2.28.1
scikit-learn==1.3.2
scipy==1.10.0
secure-smtplib==0.1.1
simpleai==0.8.3
sinchsms==1.0.4
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
telepot==12.7
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
traceback2==1.4.0
typing_extensions==4.7.0
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.12
vine==5.1.0
virtualenv==20.0.31
wcwidth==0.2.9
weasyprint==60.1
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
yarl==1.9.2
zipp==3.10.0
zopfli==0.2.3
